Added relation and constraint to basemodel
