+++ glotaran/model/util.py
@@ -3,6 +3,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from glotaran.model import model_attribute
+
 if TYPE_CHECKING:
     from typing import Any
     from typing import Callable
@@ -53,3 +55,38 @@ def wrap_func_as_method(
         return func
 
     return wrapper
+
+
+@model_attribute(
+    properties={
+        "interval": {"type": list[tuple[Any, Any]], "default": None},
+    },
+)
+class IntervalProperty:
+    """Applies a relation between clps as
+
+    :math:`source = parameter * target`.
+    """
+
+    def applies(self, index: Any) -> bool:
+        """
+        Returns true if the index is in one of the intervals.
+
+        Parameters
+        ----------
+        index :
+
+        Returns
+        -------
+        applies : bool
+
+        """
+        if self.interval is None:
+            return True
+
+        def applies(interval):
+            return interval[0] <= index <= interval[1]
+
+        if isinstance(self.interval, tuple):
+            return applies(self.interval)
+        return not any([applies(i) for i in self.interval])
